Make `BlockDiagonal` block storage generic
Before this commit, `BlockDiagonal` uses a `Vector{V}` to store the
blocks. The issue is that for trimmable binaries it is advantageous to
use stronger typing, such as `SizedVector` from `StaticArrays`.
To support such cases, we make the vector type generic, and make the
required modifications in the rest of the package.